feat: simplified and componentized main App layout
