+++ react_ui/src/components/About/index.ts
@@ -1,0 +1,2 @@
+import About from "./About";
+export default About;
